+++ __tests__/get-tokens.spec.js
@@ -3,11 +3,11 @@
  */
 require('dotenv').config();
 const request = require('supertest');
-const server = require('../server/app');
 const { expect } = require('chai');
-const seed = require('./seed');
 const sinon = require('sinon');
 const chai = require('chai');
+const seed = require('./seed');
+const server = require('../server/app');
 chai.use(require('chai-uuid'));
 
 describe('GET tokens', () => {
@@ -106,7 +106,7 @@ describe('GET tokens', () => {
   it(`GET /tokens/ should return correct offset`, async () => {
     const insertedId = (await seed.addTokenToWallet(seed.wallet.id))[0].id;
     const res = await request(server)
-      .get(`/tokens?start=1&limit=10`)
+      .get(`/tokens?offset=1&limit=10`)
       .set('treetracker-api-key', seed.apiKey)
       .set('Authorization', `Bearer ${bearerToken}`);