+++ database/migrations/20210211020241-AddDefaultIdTransactionTable.js
@@ -1,11 +1,12 @@
-'use strict';
 
-var dbm;
-var type;
-var seed;
-var fs = require('fs');
-var path = require('path');
-var Promise;
+
+let dbm;
+let type;
+let seed;
+const fs = require('fs');
+const path = require('path');
+
+let Promise;
 
 /**
   * We receive the dbmigrate dependency from dbmigrate initially.
@@ -19,11 +20,11 @@ exports.setup = function(options, seedLink) {
 };
 
 exports.up = function(db) {
-  var filePath = path.join(__dirname, 'sqls', '20210211020241-AddDefaultIdTransactionTable-up.sql');
+  const filePath = path.join(__dirname, 'sqls', '20210211020241-AddDefaultIdTransactionTable-up.sql');
   return new Promise( function( resolve, reject ) {
     fs.readFile(filePath, {encoding: 'utf-8'}, function(err,data){
       if (err) return reject(err);
-      console.log('received data: ' + data);
+      console.log(`received data: ${  data}`);
 
       resolve(data);
     });
@@ -34,11 +35,11 @@ exports.up = function(db) {
 };
 
 exports.down = function(db) {
-  var filePath = path.join(__dirname, 'sqls', '20210211020241-AddDefaultIdTransactionTable-down.sql');
+  const filePath = path.join(__dirname, 'sqls', '20210211020241-AddDefaultIdTransactionTable-down.sql');
   return new Promise( function( resolve, reject ) {
     fs.readFile(filePath, {encoding: 'utf-8'}, function(err,data){
       if (err) return reject(err);
-      console.log('received data: ' + data);
+      console.log(`received data: ${  data}`);
 
       resolve(data);
     });