+++ server/repositories/TokenRepository.spec.js
@@ -1,8 +1,9 @@
-const knex = require("../database/knex");
-const TokenRepository = require("./TokenRepository");
 const sinon = require("sinon");
 const {expect} = require("chai");
 const mockKnex = require("mock-knex");
+const TokenRepository = require("./TokenRepository");
+const knex = require("../database/knex");
+
 const tracker = mockKnex.getTracker();
 const Session = require("../models/Session");
 
@@ -26,7 +27,7 @@ describe("TokenRepository", () => {
       expect(query.sql).match(/select.*token.*transaction.*transfer_id/is);
       query.response([{id:1, token: "testUuid"}]);
     });
-    const tokens = await tokenRepository.getByTransferId("testUuid");
+    const tokens = await tokenRepository.getByTransferId('226f76cd-52b0-486b-b58a-98230696c748');
     expect(tokens).lengthOf(1);
   });