+++ server/repositories/TransactionRepository.js
@@ -1,9 +1,8 @@
-const BaseRepository = require("./BaseRepository");
-const expect = require("expect-runtime");
+const BaseRepository = require('./BaseRepository');
 
-class TransferRepository extends BaseRepository{
-  constructor(session){
-    super("transaction", session);
+class TransferRepository extends BaseRepository {
+  constructor(session) {
+    super('transaction', session);
     this._session = session;
   }
 }