+++ server/routes/trustRouter.js
@@ -1,72 +1,64 @@
 const express = require('express');
-const trustRouter = express.Router();
-const { check, validationResult } = require('express-validator');
-const assert = require("assert");
-const WalletService = require("../services/WalletService");
-const TrustService = require("../services/TrustService");
-const Wallet = require("../models/Wallet");
-const helper = require("./utils");
-const Session = require("../models/Session");
-const TrustRelationship = require("../models/TrustRelationship");
-const Joi = require("joi");
 
-trustRouter.get('/',
+const trustRouter = express.Router();
+const Joi = require('joi');
+const WalletService = require('../services/WalletService');
+const TrustService = require('../services/TrustService');
+const helper = require('./utils');
+const Session = require('../models/Session');
+const TrustRelationship = require('../models/TrustRelationship');
+
+trustRouter.get(
+  '/',
   helper.apiKeyHandler,
   helper.verifyJWTHandler,
-  helper.handlerWrapper(async (req, res, next) => {
+  helper.handlerWrapper(async (req, res) => {
     Joi.assert(
       req.query,
       Joi.object({
         state: Joi.string(),
         type: Joi.string(),
         request_type: Joi.string(),
-        start: Joi.number(),
-        limit: Joi.number().min(1).max(10000).integer(),
-      })
-    )
+        offset: Joi.number().min(0).default(0).integer(),
+        limit: Joi.number().min(1).max(1000).integer().default(1000),
+      }),
+    );
     Joi.assert(
       res.locals,
       Joi.object({
-        wallet_id: Joi.string().required()
-      })
-    )
-    const {state, type, request_type, limit, start} = req.query;
+        wallet_id: Joi.string().required(),
+      }),
+    );
+    const { state, type, request_type, limit, offset } = req.query;
     const session = new Session();
     const walletService = new WalletService(session);
     const trustService = new TrustService(session);
     const wallet = await walletService.getById(res.locals.wallet_id);
     const trust_relationships = await wallet.getTrustRelationships(
-      req.query.state,
-      req.query.type,
-      req.query.request_type,
+      state,
+      type,
+      request_type,
+      Number(offset || 0),
+      Number(limit || 0),
     );
     const subWallets = await wallet.getSubWallets();
-    for(const sw of subWallets){
+    for (const sw of subWallets) {
       const trustRelationships = await sw.getTrustRelationships(
         req.query.state,
         req.query.type,
         req.query.request_type,
       );
-      for(tr of trustRelationships){
-        if(trust_relationships.every(e => e.id !== tr.id)){
+      for (const tr of trustRelationships) {
+        if (trust_relationships.every((e) => e.id !== tr.id)) {
           trust_relationships.push(tr);
         }
       }
     }
 
-    let trust_relationships_json = [];
-    for(let t of trust_relationships){
+    const trust_relationships_json = [];
+    for (const t of trust_relationships) {
       const j = await trustService.convertToResponse(t);
       trust_relationships_json.push(j);
-    }
-
-    //filter trust_relationships json by query
-    let numStart = parseInt(start);
-    let numLimit = parseInt(limit) ? parseInt(limit) : 0; //TODO: fix this correctly by using db
-    let numBegin = numStart?numStart-1:0;
-    let numEnd = numBegin + ((numLimit != 0) ? numLimit : 1000);
-    if(numEnd != 0){
-      trust_relationships_json = trust_relationships_json.slice(numBegin, numEnd);
     }
 
     res.status(200).json({
@@ -75,32 +67,39 @@ trustRouter.get('/',
   }),
 );
 
-trustRouter.post('/',
+trustRouter.post(
+  '/',
   helper.apiKeyHandler,
   helper.verifyJWTHandler,
   helper.handlerWrapper(async (req, res) => {
     Joi.assert(
       req.body,
       Joi.object({
-        trust_request_type: Joi.string().required().valid(...Object.keys(TrustRelationship.ENTITY_TRUST_REQUEST_TYPE)),
+        trust_request_type: Joi.string()
+          .required()
+          .valid(...Object.keys(TrustRelationship.ENTITY_TRUST_REQUEST_TYPE)),
         requestee_wallet: Joi.string().required(),
-      })
+      }),
     );
     Joi.assert(
       res.locals,
       Joi.object({
-        wallet_id: Joi.string().required()
-      })
-    )
-    
+        wallet_id: Joi.string().required(),
+      }),
+    );
+
     const session = new Session();
     const walletService = new WalletService(session);
     const trustService = new TrustService(session);
     const wallet = await walletService.getById(res.locals.wallet_id);
-    const requesteeWallet = await walletService.getByName(req.body.requestee_wallet);
+    const requesteeWallet = await walletService.getByName(
+      req.body.requestee_wallet,
+    );
     let requesterWallet = wallet;
-    if(req.body.requester_wallet){
-      requesterWallet = await walletService.getByName(req.body.requester_wallet);
+    if (req.body.requester_wallet) {
+      requesterWallet = await walletService.getByName(
+        req.body.requester_wallet,
+      );
     }
 
     const trust_relationship = await wallet.requestTrustFromAWallet(
@@ -108,28 +107,31 @@ trustRouter.post('/',
       requesterWallet,
       requesteeWallet,
     );
-    const trust_relationship_json = await trustService.convertToResponse(trust_relationship);
+    const trust_relationship_json = await trustService.convertToResponse(
+      trust_relationship,
+    );
     res.status(200).json(trust_relationship_json);
-  })
+  }),
 );
 
-trustRouter.post('/:trustRelationshipId/accept',
+trustRouter.post(
+  '/:trustRelationshipId/accept',
   helper.apiKeyHandler,
   helper.verifyJWTHandler,
   helper.handlerWrapper(async (req, res) => {
     Joi.assert(
       res.locals,
       Joi.object({
-        wallet_id: Joi.string().required()
-      })
-    )
+        wallet_id: Joi.string().required(),
+      }),
+    );
     Joi.assert(
       req.params,
       Joi.object({
-        trustRelationshipId: Joi.string().required()
-      })
-    )
-    const trustRelationshipId = req.params.trustRelationshipId;
+        trustRelationshipId: Joi.string().required(),
+      }),
+    );
+    const { trustRelationshipId } = req.params;
     const session = new Session();
     const walletService = new WalletService(session);
     const trustService = new TrustService(session);
@@ -137,26 +139,27 @@ trustRouter.post('/:trustRelationshipId/accept',
     const json = await wallet.acceptTrustRequestSentToMe(trustRelationshipId);
     const json2 = await trustService.convertToResponse(json);
     res.status(200).json(json2);
-  })
+  }),
 );
 
-trustRouter.post('/:trustRelationshipId/decline',
+trustRouter.post(
+  '/:trustRelationshipId/decline',
   helper.apiKeyHandler,
   helper.verifyJWTHandler,
   helper.handlerWrapper(async (req, res) => {
     Joi.assert(
       res.locals,
       Joi.object({
-        wallet_id: Joi.string().required()
-      })
-    )
+        wallet_id: Joi.string().required(),
+      }),
+    );
     Joi.assert(
       req.params,
       Joi.object({
-        trustRelationshipId: Joi.string().required()
-      })
-    )
-    const trustRelationshipId = req.params.trustRelationshipId;
+        trustRelationshipId: Joi.string().required(),
+      }),
+    );
+    const { trustRelationshipId } = req.params;
     const session = new Session();
     const walletService = new WalletService(session);
     const trustService = new TrustService(session);
@@ -164,26 +167,27 @@ trustRouter.post('/:trustRelationshipId/decline',
     const json = await wallet.declineTrustRequestSentToMe(trustRelationshipId);
     const json2 = await trustService.convertToResponse(json);
     res.status(200).json(json2);
-  })
+  }),
 );
 
-trustRouter.delete('/:trustRelationshipId',
+trustRouter.delete(
+  '/:trustRelationshipId',
   helper.apiKeyHandler,
   helper.verifyJWTHandler,
   helper.handlerWrapper(async (req, res) => {
     Joi.assert(
       res.locals,
       Joi.object({
-        wallet_id: Joi.string().required()
-      })
-    )
+        wallet_id: Joi.string().required(),
+      }),
+    );
     Joi.assert(
       req.params,
       Joi.object({
-        trustRelationshipId: Joi.string().required()
-      })
-    )
-    const trustRelationshipId = req.params.trustRelationshipId;
+        trustRelationshipId: Joi.string().required(),
+      }),
+    );
+    const { trustRelationshipId } = req.params;
     const session = new Session();
     const walletService = new WalletService(session);
     const trustService = new TrustService(session);
@@ -191,7 +195,7 @@ trustRouter.delete('/:trustRelationshipId',
     const json = await wallet.cancelTrustRequestSentToMe(trustRelationshipId);
     const json2 = await trustService.convertToResponse(json);
     res.status(200).json(json2);
-  })
+  }),
 );
 
 module.exports = trustRouter;