+++ server/services/WalletService.js
@@ -1,8 +1,7 @@
+const { validate: uuidValidate } = require('uuid');
 const WalletRepository = require('../repositories/WalletRepository');
 const Wallet = require('../models/Wallet');
 const HttpError = require("../utils/HttpError");
-const expect = require('expect-runtime');
-const { validate: uuidValidate } = require('uuid');
 
 class WalletService {
   constructor(session){