+++ server/utils/HttpError.js
@@ -9,8 +9,8 @@ class HttpError extends Error {
   constructor(code, message, toRollback){
     super(message);
     this.code = code;
-    //set rollback flag, so the transaction of db would rollback when catch this error
-    //set default to true
+    // set rollback flag, so the transaction of db would rollback when catch this error
+    // set default to true
     this._toRollback = toRollback || true;
   }